git: remove unneeded casts
Many of these casts remain,
even though the target variable is the type it is being casted to.
We can safely remove said casts.

+++ thread-utils.c
@@ -57,7 +57,7 @@ int online_cpus(void)
 #endif /* defined(HAVE_BSD_SYSCTL) && defined(HW_NCPU) */
 
 #ifdef _SC_NPROCESSORS_ONLN
-	if ((ncpus = (long)sysconf(_SC_NPROCESSORS_ONLN)) > 0)
+	if ((ncpus = sysconf(_SC_NPROCESSORS_ONLN)) > 0)
 		return (int)ncpus;
 #endif
